requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.23
duckdb-engine>=1.1.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=2.0.0
networkx>=3.2
Faker>=21.0.0
cryptography>=41.0.0
PyJWT>=2.8.0
requests>=2.31.0
Werkzeug>=3.0.0

# Testing and linting
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.3.0
black>=23.12.0
flake8>=6.1.0
